+++ lib/src/domain/repositories/miners_repository.dart
@@ -6,7 +6,7 @@ abstract class MinersRepository {
   Future<Miner> getMinerData(String minerId);
 
   Future<CurrentStats> updateMinerCurrentStats(String minerId);
-  Future<int> updateMinerMinPayout(String minerId);
+  Future<double> updateMinerMinPayout(String minerId);
   Future<List<MinerHistory>> updateMinerHistory(String minerId);
   Future<List<MinerPayout>> updateMinerPayouts(String minerId);
 }
